requirements.txt
APScheduler==3.10.4
attrs==21.2.0
Automat==20.2.0
awscli==1.32.27
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
botocore==1.34.27
bs4==0.0.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
cloud-init==23.3.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.16
ec2-hibinit-agent==1.0.0
elastic-transport==8.12.0
elasticsearch==8.12.0
Flask==3.0.1
hibagent==1.0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.4
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
pexpect==4.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==3.0.1
PyPDF3==1.0.6
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.10.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.5.6
soupsieve==2.5
ssh-import-id==5.11
systemd-python==234
tqdm==4.66.1
Twisted==22.1.0
tzlocal==5.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.0.1
zipp==1.0.0
zope.interface==5.4.0
